DeStruct: set local file
