+++ aspnet-core/src/FMS.Application/FMS.Application.csproj
@@ -14,7 +14,4 @@
   <ItemGroup>
     <ProjectReference Include="..\FMS.Core\FMS.Core.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Orders\Dto\" />
-  </ItemGroup>
 </Project>
