+++ aspnet-core/src/FMS.Application/Orders/Dto/PagedOrderResultRequestDto.cs
@@ -1,6 +1,6 @@
 ﻿using Abp.Application.Services.Dto;
 
-namespace FMS.Orders
+namespace FMS.Orders.Dto
 {
     public class PagedOrderResultRequestDto : PagedResultRequestDto
     {
