+++ aspnet-core/src/FMS.Application/Orders/OrderAppService.cs
@@ -1,11 +1,11 @@
 ﻿using Abp.Application.Services;
-using FMS.Orders;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Abp.Domain.Repositories;
+using FMS.Orders.Dto;
 
 namespace FMS.Orders
 {
